projectFinalv3

+++ PROYECTOFINAL/BACKEND_NODE/src/controllers/carts.controller.js
@@ -1,5 +1,7 @@
+import { ticketModel } from "../DAL/mongoDB/models/tickets.model.js";
 import { cartsService } from "../services/cart.service.js";
 import { productsService } from "../services/products.service.js";
+import { ticketService } from "../services/tickets.service.js";
 import { usersService} from "../services/users.service.js"
 
 class CartsController {
@@ -15,7 +17,6 @@ class CartsController {
   
   async findOneCart(req, res) {
     const { idcart } = req.params;
-    console.log(idcart);
     try {
       const response = await cartsService.findOneCart( idcart)
       res.status(200).json({ message: "Cart", response });
@@ -26,7 +27,6 @@ class CartsController {
 
   async findCartById(req, res) {
     const { idcart } = req.params;
-    console.log(idcart);
     try {
       const response = await cartsService.findCartById(idcart)
       res.status(200).json({ message: "Cart", response });
@@ -51,8 +51,6 @@ class CartsController {
   }
 
   async updatedCart(req, res) {
-    console.log("test to updated");
-    console.log(req.body);
     const{cartId} =req.body
     try {    
       
@@ -67,62 +65,93 @@ class CartsController {
   }
   async purchaseCart(req, res) {
     const id = req.params.id; // obtengo id del carrito mediante params 
-    console.log(id);
     try {
       const cart = await cartsService.findCartById(id);// busco el carrito mediante cartService
       console.log("carrito recibido",cart);
+      const userId = cart.userId
+      
+      console.log("id del usuario" , userId);// purcharser
+
+
       if (!cart) {
         return res.status(404).json({ error: "Carrito no encontrado" });
-      }  
+      } 
+      const ticket = {
+        purchaser: userId , // Set this later
+        products: [],
+        amount: 0,
+      };
     for (const item of cart.products) {// itero los productos de mi carrito con un for of
 
         const idProduct = item.id_prod._id // guardo en newobjecto el string del id product .. con JSON.strigfy no me permitia por tamaño de bytes
         console.log("id del Producto",idProduct);
         const product = await productsService.findOneProducts(idProduct);// busco el producto 
-        console.log("producto a comprar",product);
+        console.log("producto a comprar",product.title);
         console.log("cantidad",item.cantidad);
         const requestedQuantity = item.cantidad;
-        console.log(product.stock);
+        console.log("stock del producto",product.stock);
+        
         if (product.stock >= requestedQuantity) {// verifico si tengo stock
             product.stock-=requestedQuantity // operador -- para descontar stock
-            console.log(product.stock)
-            console.log(product.id)
+            
             const productUpdated=await productsService.updatedProduct(product.id,product)// actualizo base de datos de product stock
-            console.log("updated",productUpdated);
-            console.log(`Hay suficiente cantidad de ${product.title} en stock.`);
+            
+            const productPrice= product.price
+            const subtotal = parseFloat((requestedQuantity * productPrice).toFixed(2));
+
+        const ticketProduct = {
+          id_prod: product._id,
+          cantidad: requestedQuantity,
+          subtotal: subtotal, // Include the subtotal for the product
+        };
+            
+            ticket.products.push(ticketProduct);
+            
         } else {
           console.log(`No hay suficiente cantidad de ${product.title} en stock.`);
         }
       }
 
+
+      // Calculate the total amount based on purchased products
+      const total = parseFloat(
+        ticket.products.reduce(
+          (accumulatedTotal, product) => accumulatedTotal + product.subtotal,
+          0
+        ).toFixed(2) // Calculate and parse total as a number with 2 decimal places
+      );
+  
+      ticket.amount = total; // Set the total amount in the ticket
+  
+      console.log("Ticket", ticket);
+
+      
+      // Save the ticket to the database
+      const createdTicket = await ticketService.createTicket(ticket)
+      
+      if (!createdTicket) {
+        throw new Error("Ticket creation failed.");
+      }
+
       await cartsService.deleteCart(id);
 
-      return res.json({ message: "Compra realizada exitosamente", cart });
+      return res.json({ message: "Compra realizada exitosamente", createdTicket });
     } catch (error) {
       throw new Error("error cart controller.");
     }
   }
 
-  async deleteCartProduct(req, res) {
+  async deleteProductFromCart(req, res) {
     const { cartId, productId } = req.body;
-    console.log("1");
-    console.log(typeof(productId));
-    // const productIdJson = JSON.parse(productId)
     try {
       const cart = await cartsService.findCartById(cartId)
       console.log(cart.products);
       if (!cart) {
         return res.status(404).json({ message: 'Carrito no encontrado' });
       }
-      console.log(2);
-      // cart.products = cart.products.filter(product => JSON.stringify(product.id_prod) !== productId);
       cart.products = cart.products.filter(product => {
-        console.log('Producto en proceso:', product.id_prod._id);
         return product.id_prod._id.toString() !== productId;
       });
-      console.log("afterupdates");
-      console.log(cart.products);
-      
       const updatedCart = await cart.save();
 
       res.status(200).json({ message: 'Producto eliminado del carrito', cart: updatedCart });
